Moves logger back out to the Terminus namespace
